+++ Clicked/Definitions.lua
@@ -72,7 +72,7 @@
 --- @field public auraName string|integer
 --- @field public interrupt boolean
 --- @field public executionOrder integer
---- @field public convertValueToId boolean
+--- @field public spellMaxRank boolean
 --- @field public startAutoAttack boolean
 --- @field public startPetAttack boolean
 --- @field public cancelQueuedSpell boolean